Split up SSR renderers into modules
